requirements.txt
langchain==0.1.4
langchain_community==0.0.16
langchain_core==0.1.17
python-dotenv==1.0.1
streamlit==1.30.0
chromadb==0.3.29
beautifulsoup4==4.12.2
google-generativeai
langchain_google_genai
faiss-cpu==1.8.1
